requirements.txt
alpaca-py==0.8.2
certifi==2023.5.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.4
holidays==0.25
idna==3.4
importlib-resources==5.12.0
kiwisolver==1.4.4
korean-lunar-calendar==0.3.1
matplotlib==3.7.1
msgpack==1.0.5
numpy==1.24.3
packaging==23.1
pandas==1.5.3
Pillow==9.5.0
pydantic==1.10.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
six==1.16.0
sseclient-py==1.7.2
typing_extensions==4.6.2
urllib3==2.0.2
websockets==10.4
zipp==3.15.0
